pyproject.toml
[tool.poetry]
name = "stay"
version = "0.0.1"
description = "stay"
authors = ["Chisu Park <soosin-spresto@spresto.net>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.22.0"
arrow = "^0.15.4"
authlib = "^0.13"
email_validator = "^1.0.5"
python-multipart = "^0.0.5"
pydantic = "^1.8.1"
gunicorn = "^20.0.4"
Pillow = "^8.3.1"
aiofiles = "^0.7.0"
boto3 = "^1.20.21"
Django = "^4.0.1"
fastapi-versioning = "^0.10.0"
fastapi = "^0.64.0"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.1"
pytz = "^2023.3"



[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.7.9"
mypy = "^0.761"
pytest = "^5.3.4"
pytest-django = "^4.4.0"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py38']
line-length = 120
skip-string-normalization = true
exclude = '(migrations)'
